Option to choose push notification level
Settings to choose the amount of information that's ported to slack.
Minimal -> Shows only the first line that's default for all the messages.
Compact -> Minimal and a simple list of the commits (hash, first line of the commit message and user)
Verbose -> The original message.

The compact message is inspired by the github bot for slack.